Generate rewriting lemmas for data types
If no rewrite lemma is specified in a 'rewrite' construct, and one is
available for the types in the equality, apply that.
Rewriting lemmas may fail to generate in some cases (I haven't yet
worked out exactly what the criteria are, and if we can fix it...) in
which case it fails silently, at least until the rewrite is applied.

The result is that, for example, rewrites on equalities on vectors and
other similar types can be applied without doing anything special.